pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-parser"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Async DB
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # PDF extraction
    "pymupdf>=1.24.0",
    "pdfplumber>=0.11.0",

    # HTTP client (for Mistral OCR)
    "httpx>=0.27.0",

    # LLM extraction
    "google-genai>=1.0.0",

    # Validation / serialisation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Logging
    "structlog>=24.2.0",

    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",   # TestClient
    "factory-boy>=3.3.0",
    "locust>=2.29.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
